EdgeDevice/requirements.txt
pillow
requests
flask